+++ sdk/resourcemanager/datadog/armdatadog/tagrules_client_example_test.go
@@ -55,7 +55,7 @@ func ExampleTagRulesClient_CreateOrUpdate() {
 	_ = res
 	// If the HTTP response code is 200 as defined in example definition, your response structure would look as follows. Please pay attention that all the values in the output are fake values for just demo purposes.
 	// res = armdatadog.TagRulesClientCreateOrUpdateResponse{
-	// 	MonitoringTagRules: &armdatadog.MonitoringTagRules{
+	// 	MonitoringTagRules: armdatadog.MonitoringTagRules{
 	// 		Name: to.Ptr("default"),
 	// 		Type: to.Ptr("Microsoft.Datadog/monitors/tagRules"),
 	// 		ID: to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.Datadog/monitors/myMonitor/tagRules/default"),
@@ -106,7 +106,7 @@ func ExampleTagRulesClient_Get() {
 	_ = res
 	// If the HTTP response code is 200 as defined in example definition, your response structure would look as follows. Please pay attention that all the values in the output are fake values for just demo purposes.
 	// res = armdatadog.TagRulesClientGetResponse{
-	// 	MonitoringTagRules: &armdatadog.MonitoringTagRules{
+	// 	MonitoringTagRules: armdatadog.MonitoringTagRules{
 	// 		Name: to.Ptr("default"),
 	// 		Type: to.Ptr("Microsoft.Datadog/monitors/tagRules"),
 	// 		ID: to.Ptr("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/myResourceGroup/providers/Microsoft.Datadog/monitors/myMonitor/tagRules/default"),